+++ CHANGES.md
@@ -8,6 +8,7 @@
   an array
 - add method override support in form post for DELETE and PUT by setting the
   hidden form field "_METHOD" to either "PUT" or "DELETE"
+- implement default route support using `Service::setDefaultRoute`
 
 ## 0.5.1
 - fix bug with registerBeforeEachMatchPlugin where skipping a plugin didn't 
@@ -63,7 +64,6 @@ Other changes:
 - add `getBaseUri()` method to `Uri` object
 - normalize Uri by stripping ports if they are the default ports
   for HTTP (80) and HTTPS (443)
-- implement default route support using `Service::setDefaultRoute`
 
 ## 0.4.7
 - support multiple request methods per match line