+++ CHANGES.md
@@ -1,5 +1,8 @@
 # Release History
 
+## 0.5.3
+- fix `MellonAuthentication` plugin
+
 ## 0.5.2
 - add `MellonAuthentication` plugin for SAML support to authenticate users
 - `BasicAuthentication` plugin now returns a UserInfo object so it becomes